Move bigint tests into dialects directory
